Change: Rename /me to /me/get
