+++ sos21-api-server/schema/api.yml
@@ -14,9 +14,9 @@ servers:
   - url: "http://localhost:4010"
     description: Local server (with validation proxy)
 paths:
-  /me:
+  /me/get:
     get:
-      summary: me
+      summary: me/get
       tags:
         - me
         - user
@@ -39,7 +39,7 @@ paths:
           $ref: "#/components/responses/Forbidden"
         "500":
           $ref: "#/components/responses/InternalServerError"
-      operationId: me
+      operationId: me/get
       description: ログインしているユーザーの情報を返します。
       parameters: []
   /signup:
